:construction: Working on Language System (Read commit desc.)
This commit will not work due to the many errors that you will receive when trying to launch the bot. All this is due to the transition to a new language system and the need for a complete rewrite of the code of all commands and events of the bot.